feat: add fields and guards to tours

+++ src/tours/tours.service.ts
@@ -1,37 +1,71 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { TourEntity } from './entities/tours.entity';
+import { CreateTourDto } from './dto/create-tour.dto';
+import { UpdateTourDto } from './dto/update-tour.dto';
+import { UsersService } from 'src/users/users.service';
 
 @Injectable()
 export class ToursService {
   constructor(
     @InjectRepository(TourEntity)
     private toursRepository: Repository<TourEntity>,
+    private usersService: UsersService,
   ) {}
 
   async findAll(): Promise<TourEntity[]> {
-    return await this.toursRepository.find();
+    return await this.toursRepository.find({
+      relations: {
+        author: true,
+      },
+    });
   }
 
-  async create(name: string, description: string): Promise<TourEntity> {
-    const newTour = this.toursRepository.create({ name, description });
-    return await this.toursRepository.save(newTour);
+  async create(
+    createTourData: CreateTourDto,
+    author_uuid: string,
+  ): Promise<TourEntity> {
+    const author = await this.usersService.findById(author_uuid);
+
+    return await this.toursRepository.save({
+      author: author,
+      ...createTourData,
+    });
   }
-  async delete(uuid: string): Promise<void> {
-    const tourToDelete = await this.toursRepository.findOne({ where: { uuid: uuid } });
+
+  async delete(tourId: string, userId: string): Promise<void> {
+    const tourToDelete = await this.toursRepository.findOneBy({ uuid: tourId });
+
+    if (!tourId) {
+      throw new NotFoundException('Tour not found');
+    }
+    if (tourToDelete.author.id !== userId) {
+      throw new ForbiddenException('Provided user is not author of this tour');
+    }
+
     //TODO: error handling
     await this.toursRepository.remove(tourToDelete);
+  }
 
-  }
-  async update(uuid: string, newName: string, newDescription: string): Promise<TourEntity> {
-    const tourToUpdate = await this.toursRepository.findOne({ where: { uuid: uuid } });
+  async patch(
+    tourId: string,
+    newTourData: UpdateTourDto,
+    userId: string,
+  ): Promise<TourEntity> {
+    const tourToUpdate = await this.toursRepository.findOneBy({ uuid: tourId });
+
     if (!tourToUpdate) {
-      throw new Error(`Tour with UUID ${uuid} not found`);
+      throw new NotFoundException(`Tour with UUID ${tourId} not found`);
     }
-    tourToUpdate.name = newName;
-    tourToUpdate.description = newDescription;
-    return this.toursRepository.save(tourToUpdate);
+    if (tourToUpdate.author.id !== userId) {
+      throw new ForbiddenException('Provided user is not author of this tour');
+    }
+
+    return this.toursRepository.save({ uuid: tourId, ...newTourData });
   }
-  
 }
